+++ ui/page/collection/internal/collectionPublishForm/internal/collectionGeneralTab/internal/additionalOptions/view.jsx
