+++ ui/page/collection/internal/collectionPublishForm/internal/collectionGeneralTab/view.jsx
